apps/api/requirements.txt
camelot-py[base]===0.10.1
requests===2.27.1
